+++ src/client/core/chatWindow.h
