+++ src/client/core/mainWindow.h
@@ -1,0 +1,30 @@
+#ifndef MAINWINDOW_H
+#define MAINWINDOW_H
+
+#include <QMainWindow>
+#include "loginSystem.h"
+#include "chatWindow.h"
+
+QT_BEGIN_NAMESPACE
+namespace Ui { class MainWindow; }
+QT_END_NAMESPACE
+
+class MainWindow : public QMainWindow {
+Q_OBJECT
+
+public:
+    MainWindow(QWidget *parent = nullptr);
+
+    ~MainWindow();
+
+private:
+    Ui::MainWindow *ui;
+    ChatWindow *chat;
+
+private slots:
+
+    void open();
+    void on_pushButton_clicked();
+};
+
+#endif // MAINWINDOW_H
