--- src/client/mainWindow.h
@@ -1,30 +1,0 @@
-#ifndef MAINWINDOW_H
-#define MAINWINDOW_H
-
-#include <QMainWindow>
-#include "loginSystem.h"
-#include "chatWindow.h"
-
-QT_BEGIN_NAMESPACE
-namespace Ui { class MainWindow; }
-QT_END_NAMESPACE
-
-class MainWindow : public QMainWindow {
-Q_OBJECT
-
-public:
-    MainWindow(QWidget *parent = nullptr);
-
-    ~MainWindow();
-
-private:
-    Ui::MainWindow *ui;
-    ChatWindow *chat;
-
-private slots:
-
-    void open(); // an openning the main window function
-    void on_pushButton_clicked(); // event which does pushing the button
-};
-
-#endif // MAINWINDOW_H
